+++ src/services/color/visualization_colors.ts
@@ -1,7 +1,7 @@
 // Array of color-blind safe colors to use in visualizations or other
 // spots that need a large range of varied, qualitative colors.
-import { palettes } from '../../services/color/eui_palettes';
+import { euiPaletteColorBlind } from '../../services/color/eui_palettes';
 
-export const VISUALIZATION_COLORS = palettes.euiPaletteColorBlind();
+export const VISUALIZATION_COLORS = euiPaletteColorBlind();
 
 export const DEFAULT_VISUALIZATION_COLOR = VISUALIZATION_COLORS[1];